Move method to other CharSequence extensions
